analyzer_service - Copy/requirements.txt
numpy
paho-mqtt
brainflow